requirements.txt
psycopg2-binary>=2.8.6
Django==3.1
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-map-widgets==0.3.0
django-phonenumber-field==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
phonenumbers==8.12.7
Pillow==7.2.0
python-decouple==3.3
django-google-maps==0.12.2
django-admin-thumbnails==0.2.5
djangorestframework-gis==0.15
gunicorn>=19.7.1,<20
whitenoise>=4.1.2
django-admin-sortable2>=0.7.7